requirements.txt
flask==3.0.0
openai==1.51.2
anthropic==0.34.2
google-generativeai==0.8.3
slack-sdk==3.33.1
python-dotenv==1.0.1
gunicorn==21.2.0